LPS-115615 Persist native object when a new native object service is registered
